BACKPORT: arm64: Kconfig: Enable NODES_SPAN_OTHER_NODES config for NUMA
Some NUMA nodes have memory ranges that span other nodes.
Even though a pfn is valid and between a node's start and end pfns,
it may not reside on that node.

This patch enables NODES_SPAN_OTHER_NODES config for NUMA to support
this type of NUMA layout.

This patch is in open source review progress and is backported from:
https://www.spinics.net/lists/arm-kernel/msg737306.html

cherry-pick from:
https://github.com/AmpereComputing/ampere-centos-kernel/commit/9db4e45

Signed-off-by: Hoan Tran <hoan.tran@amperecomputing.com>
Signed-off-by: Khuong Dinh <khuong.dinh@amperecomputing.com>