NOUPSTREAM: firmware: Support Ampere Altra CPUECTLR_EL1 config driver
Based on the spec, CPUECTLR_EL1 cannot configure on the fly

" Accessing the CPUECTLR_EL1
The CPU Extended Control Register can be written only when the system is
idle. Arm recommends that you write to this register after a powerup
reset before the MMU is enabled. "

To support changing CPUECTLR_EL1 without the system reboot,
here is the sequence:
- Create a non-secure "Shared Memory" to store the CPUECTLR_EL1
  value
- From kernel, the user configures the desired value into the "Shared
  Memory" region from a new kernel driver
- From kernel, the user takes each core offline then online until finish
  all the cores
         + When the core is online, at first, it jumps into ATF (EL3)
         + At ATF (EL3), core read the "Shared Memory" to get the new
CPUECTLR_EL1 config value
         + Core configures the new value into CPUECTLR_EL1 register
         + Core jumps into kernel and enable MMU.
- All the cores run with the new CPUECTLR_EL1 value.

This patch adds the new kernel driver to configure the NS-DRAM
shared memory.

It's for the customer and internal testing purpose.

Signed-off-by: Hoan Tran <hoan.tran@amperecomputing.com>
Signed-off-by: Bobo <lmw.bobo@gmail.com>